rag_system/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.6
celery[redis]==5.3.6

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Document parsing
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23

# Tokenization
tiktoken==0.5.2

# Embeddings and Vector Database
google-generativeai==0.3.2
pinecone-client==3.0.0

# Logging
python-json-logger==2.0.7

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
numpy==1.26.3
